src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_FinitePlaneTest.java
/*
 * Copyright 2020 Andy Turner, University of Leeds.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package uk.ac.leeds.ccg.v3d.geometry;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

/**
 *
 * @author geoagdt
 */
public class V3D_FinitePlaneTest {
    
    public V3D_FinitePlaneTest() {
    }
    
    @BeforeAll
    public static void setUpClass() {
    }
    
    @AfterAll
    public static void tearDownClass() {
    }
    
    @BeforeEach
    public void setUp() {
    }
    
    @AfterEach
    public void tearDown() {
    }

    /**
     * Test of getEnvelope3D method, of class V3D_FinitePlane.
     */
    @Test
    public void testGetEnvelope3D() {
        System.out.println("getEnvelope3D");
        V3D_FinitePlane instance = null;
        V3D_Envelope expResult = null;
        V3D_Envelope result = instance.getEnvelope3D();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getPlane method, of class V3D_FinitePlane.
     */
    @Test
    public void testGetPlane() {
        System.out.println("getPlane");
        V3D_FinitePlane instance = null;
        V3D_Plane expResult = null;
        V3D_Plane result = instance.getPlane();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of isIntersectedBy method, of class V3D_FinitePlane.
     */
    @Test
    public void testIsIntersectedBy() {
        System.out.println("isIntersectedBy");
        V3D_Point pt = null;
        V3D_FinitePlane instance = null;
        boolean expResult = false;
        boolean result = instance.isIntersectedBy(pt);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getIntersection method, of class V3D_FinitePlane.
     */
    @Test
    public void testGetIntersection() {
        System.out.println("getIntersection");
        V3D_Line li = null;
        V3D_FinitePlane instance = null;
        V3D_Geometry expResult = null;
        V3D_Geometry result = instance.getIntersection(li);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
